refactor(notification): settingsファイル移動
