+++ lambda/notification/src/settings.py
